Moved modules to designated folder
