+++ src/modules/logFunctions.ts
